+++ classes/Gems/Condition/ConditonLoadException.php
@@ -1,0 +1,25 @@
+<?php
+                
+/**
+ *
+ * @package    Gem
+ * @subpackage Condition
+ * @author     Matijs de Jong <mjong@magnafacta.nl>
+ * @copyright  Copyright (c) 2020, Erasmus MC and MagnaFacta B.V.
+ * @license    No free license, do not copy
+ */
+
+namespace Gems\Condition;
+
+/**
+ *
+ *
+ * @package    Gem
+ * @subpackage Condition
+ * @license    No free license, do not copy
+ * @since      Class available since version 1.8.8
+ */
+class ConditonLoadException extends \Gems_Exception_Coding
+{
+
+}