+++ classes/Gems/Event/Application/EventRegistry.php
@@ -4,6 +4,15 @@
 namespace Gems\Event\Application;
 
 
+/**
+ * Central registry for available event types, code not executed
+ *
+ * @package    Gems
+ * @subpackage Event
+ * @copyright  Copyright (c) 2020 Erasmus MC & Equipe Zorgbedrijven
+ * @license    New BSD License
+ * @since      Class available since version 1.8.8
+ */
 class EventRegistry
 {
     public static $events = [
@@ -13,9 +22,12 @@ class EventRegistry
         MenuAdd::Name => MenuAdd::class,
         // Set the correct directory for the current Controller
         SetFrontControllerDirectory::NAME => SetFrontControllerDirectory::class,
+
+        'gems.tracker.conditions.get' => TranslatableNamedArrayEvent::class,
+
         // Get an array of available Track Field Types for the FieldMaintenanceModel
         'gems.tracker.fieldtypes.get' => TranslatableNamedArrayEvent::class,
         // Get an array of available Track Field Dependencies for the FieldMaintenanceModel
-        'gems.tracker.fieldtypes.get' => NamedArrayEvent::class,
+        'gems.tracker.fielddependencies.get' => NamedArrayEvent::class,
     ];
 }